+++ src/app/configs/routesConfig.js
@@ -3,12 +3,10 @@ import FuseLoading from '@fuse/core/FuseLoading';
 import { Navigate } from 'react-router-dom';
 import settingsConfig from 'app/configs/settingsConfig';
 import SignInConfig from '../main/pages/auth/sign-in/SignInConfig';
-import SignUpConfig from '../main/pages/auth/sign-up/SignUpConfig';
 import SignOutConfig from '../main/pages/auth/sign-out/SignOutConfig';
 import ForgotPasswordConfig from '../main/pages/auth/password/forgot/ForgotPasswordConfig';
 import ResetPasswordConfig from '../main/pages/auth/password/reset/ResetPasswordConfig';
-import EmailSentConfig from '../main/pages/auth/email/sent/EmailSentConfig';
-import ConfirmEmailConfig from '../main/pages/auth/email/confirm/ConfirmEmailConfig';
+import ConcludeConfig from '../main/pages/auth/conclude/ConcludeConfig';
 import Error404Page from '../main/404/Error404Page';
 import ExampleConfig from '../main/example/ExampleConfig';
 
@@ -16,10 +14,10 @@ const routeConfigs = [
   ExampleConfig,
   SignOutConfig,
   SignInConfig,
-  SignUpConfig,
   EmailSentConfig,
   ForgotPasswordConfig,
   ResetPasswordConfig,
+  ConcludeConfig,
   ConfirmEmailConfig
 ];
