--- src/app/main/pages/auth/sign-up/SignUpConfig.js
@@ -1,35 +1,0 @@
-import SignUpPage from './SignUpPage';
-import authRoles from '../../../../auth/authRoles';
-
-const SignUpConfig = {
-  settings: {
-    layout: {
-      config: {
-        navbar: {
-          display: false,
-        },
-        toolbar: {
-          display: false,
-        },
-        footer: {
-          display: false,
-        },
-        leftSidePanel: {
-          display: false,
-        },
-        rightSidePanel: {
-          display: false,
-        },
-      },
-    },
-  },
-  auth: authRoles.onlyGuest,
-  routes: [
-    {
-      path: 'sign-up',
-      element: <SignUpPage />,
-    },
-  ],
-};
-
-export default SignUpConfig;
